move KNOWN_LINK_PATTERNS back to AssistantRuleBook for now
